requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
alembic
pydantic
python-dotenv
psycopg2-binary
gunicorn
python-dateutil